fix(settings): validate URL writes + stop swallowing save failures
Two medium findings from the 2026-06-05 review, both in the shared
`ui::settings_form::save_settings` (used by auth_ui, legalpages, userportal, …):

- M16: branding/`_URL` writes bypassed the SSRF/HTTPS validation the admin
  variables page enforces — an admin could set a private-IP/non-HTTPS URL one
  surface rejects. Now every `InputType::Url` value is validated (via
  `var.is_url()`, the declared rule) before write.
- M24: `config::set` failures were logged and then reported as
  "Settings saved" (HTTP 200). Now the first write failure surfaces as a real
  error response; htmx clients branch on the status.

To share one validator across both write surfaces, `validate_url_value` moved to
`crate::util` (the shared layer both `ui` and `blocks` already depend on);
`blocks::admin::ops` re-exports it so the admin and settings-form paths can't
diverge on what a URL value may be. URL values are validated up front so one bad
URL can't leave a half-applied save.

Tests: SSRF URL rejected, failed `config::set` surfaced as error, plus the moved
validate_url_value unit test. Full solobase-core lib suite green (774).
